refactor series indexing
